Fix not having session first

+++ tugas_2/comment/views.py
@@ -18,28 +18,31 @@ class commentMethod :
 		    body = json.loads(body_unicode)
 		    comment = body['comment']
 		    url = "https://oauth.infralabs.cs.ui.ac.id/oauth/resource"
-		    access_token = request.session['access_token']
-		    headers = {"Authorization": "Bearer " + access_token}
-		    r = requests.get(url, headers=headers)
-		    if r.status_code != 401 :
-		    	search_user = list(User.objects.filter(userId=r.json()['user_id']))
-		    	if len(search_user) != 0 :
-				    newComment = Comment(comment=comment, createdBy=search_user[0].displayName, createdAt=datetime.now(), updatedAt=datetime.now())
-				    newComment.save()
-				    return JsonResponse({
-				    	"status" : "ok", 
-				    	"data": {
-					    	"id" : newComment.id, 
-					    	"comment" : newComment.comment, 
-					    	"createdBy": newComment.createdBy, 
-					    	"createdAt": newComment.createdAt, 
-					    	"updatedAt": newComment.updatedAt
-				    	}
-				    })
-		    	else :
-			    	return JsonResponse({"status" : "error", "description": "You haven't registered"})
+		    if 'access_token' in request.session:
+			    access_token = request.session['access_token']
+			    headers = {"Authorization": "Bearer " + access_token}
+			    r = requests.get(url, headers=headers)
+			    if r.status_code != 401 :
+			    	search_user = list(User.objects.filter(userId=r.json()['user_id']))
+			    	if len(search_user) != 0 :
+					    newComment = Comment(comment=comment, createdBy=search_user[0].displayName, createdAt=datetime.now(), updatedAt=datetime.now())
+					    newComment.save()
+					    return JsonResponse({
+					    	"status" : "ok", 
+					    	"data": {
+						    	"id" : newComment.id, 
+						    	"comment" : newComment.comment, 
+						    	"createdBy": newComment.createdBy, 
+						    	"createdAt": newComment.createdAt, 
+						    	"updatedAt": newComment.updatedAt
+					    	}
+					    })
+			    	else :
+				    	return JsonResponse({"status" : "error", "description": "You haven't registered"})
+			    else :
+			    	return JsonResponse({"status" : "error", "description": "Unauthorized please login first"}, status=401)
 		    else :
-		    	return JsonResponse({"status" : "error", "description": "Unauthorized"}, status=401)
+		    	return JsonResponse({"status" : "error", "description": "Unauthorized please login first"}, status=401)
 	  	else :
 	  		return JsonResponse({"status" : "error", "description": "Method not allowed"})
 	    
@@ -50,25 +53,28 @@ class commentMethod :
 	  		body = json.loads(body_unicode)
 	  		id = body['id']
 	  		url = "https://oauth.infralabs.cs.ui.ac.id/oauth/resource"
-	  		access_token = request.session['access_token']
-	  		headers = {"Authorization": "Bearer " + access_token}
-	  		r = requests.get(url, headers=headers)
-	  		if r.status_code != 401 :
-	  			search_user = list(User.objects.filter(userId=r.json()['user_id']))
-	  			if len(search_user) != 0 :
-	  				comment_list = list(Comment.objects.filter(id=id))
-	  				if len(comment_list) != 0 :
-	  					if search_user[0].displayName == comment_list[0].createdBy :
-	  						Comment.objects.filter(id=id).delete()
-	  						return JsonResponse({"status" : "ok"})
-	  					else :
-	  						return JsonResponse({"status" : "error", "description": "You're not authorize to delete others comment"})
-	  				else :
-	  					return JsonResponse({"status" : "error", "description": "Id Not Found"})
-	  			else :
-	  				return JsonResponse({"status" : "error", "description": "You haven't registered"})
+	  		if 'access_token' in request.session:
+		  		access_token = request.session['access_token']
+		  		headers = {"Authorization": "Bearer " + access_token}
+		  		r = requests.get(url, headers=headers)
+		  		if r.status_code != 401 :
+		  			search_user = list(User.objects.filter(userId=r.json()['user_id']))
+		  			if len(search_user) != 0 :
+		  				comment_list = list(Comment.objects.filter(id=id))
+		  				if len(comment_list) != 0 :
+		  					if search_user[0].displayName == comment_list[0].createdBy :
+		  						Comment.objects.filter(id=id).delete()
+		  						return JsonResponse({"status" : "ok"})
+		  					else :
+		  						return JsonResponse({"status" : "error", "description": "You're not authorize to delete others comment"})
+		  				else :
+		  					return JsonResponse({"status" : "error", "description": "Id Not Found"})
+		  			else :
+		  				return JsonResponse({"status" : "error", "description": "You haven't registered"})
+		  		else :
+		  			return JsonResponse({"status" : "error", "description": "Unauthorized please login first"}, status=401)
 	  		else :
-	  			return JsonResponse({"status" : "error", "description": "Unauthorized"}, status=401)
+	  			return JsonResponse({"status" : "error", "description": "Unauthorized please login first"}, status=401)
 	  	else :
 	  		return JsonResponse({"status" : "error", "description": "Method not allowed"})
 	    
@@ -79,37 +85,40 @@ class commentMethod :
 		    body = json.loads(body_unicode)
 		    id = body['id']
 		    url = "https://oauth.infralabs.cs.ui.ac.id/oauth/resource"
-		    access_token = request.session['access_token']
-		    headers = {"Authorization": "Bearer " + access_token}
-		    r = requests.get(url, headers=headers)
-		    if r.status_code != 401 :
-		    	search_user = list(User.objects.filter(userId=r.json()['user_id']))
-		    	if len(search_user) != 0 :
-		    		comment_list = list(Comment.objects.filter(id=id))
-		    		if (len(comment_list) != 0) :
-			    		if search_user[0].displayName == comment_list[0].createdBy :
-			    			newComment = body['comment']
-		    				comment_list[0].comment = newComment
-		    				comment_list[0].updatedAt = datetime.now()
-		    				comment_list[0].save()
-		    				return JsonResponse({
-					    		"status" : "ok", 
-					    		"data": {
-					    			"id" : comment_list[0].id, 
-									"comment" : comment_list[0].comment, 
-									"createdBy": comment_list[0].createdBy, 
-									"createdAt": comment_list[0].createdAt, 
-								 	"updatedAt": comment_list[0].updatedAt
-								}
-							})
-			    		else :
-			    			return JsonResponse({"status" : "error", "description": "You're not authorize to update others comment"})
-	    			else :
-	    				return JsonResponse({"status" : "error", "description": "Id Not Found"})
-		    	else :
-		    		return JsonResponse({"status" : "error", "description": "You haven't registered"})
+		    if 'access_token' in request.session:
+			    access_token = request.session['access_token']
+			    headers = {"Authorization": "Bearer " + access_token}
+			    r = requests.get(url, headers=headers)
+			    if r.status_code != 401 :
+			    	search_user = list(User.objects.filter(userId=r.json()['user_id']))
+			    	if len(search_user) != 0 :
+			    		comment_list = list(Comment.objects.filter(id=id))
+			    		if (len(comment_list) != 0) :
+				    		if search_user[0].displayName == comment_list[0].createdBy :
+				    			newComment = body['comment']
+			    				comment_list[0].comment = newComment
+			    				comment_list[0].updatedAt = datetime.now()
+			    				comment_list[0].save()
+			    				return JsonResponse({
+						    		"status" : "ok", 
+						    		"data": {
+						    			"id" : comment_list[0].id, 
+										"comment" : comment_list[0].comment, 
+										"createdBy": comment_list[0].createdBy, 
+										"createdAt": comment_list[0].createdAt, 
+									 	"updatedAt": comment_list[0].updatedAt
+									}
+								})
+				    		else :
+				    			return JsonResponse({"status" : "error", "description": "You're not authorize to update others comment"})
+		    			else :
+		    				return JsonResponse({"status" : "error", "description": "Id Not Found"})
+			    	else :
+			    		return JsonResponse({"status" : "error", "description": "You haven't registered"})
+			    else :
+			    	return JsonResponse({"status" : "error", "description": "Unauthorized please login first"}, status=401)
 		    else :
-		    	return JsonResponse({"status" : "error", "description": "Unauthorized"}, status=401)
+		    	return JsonResponse({"status" : "error", "description": "Unauthorized please login first"}, status=401)
 	  	else :
 	  		return JsonResponse({"status" : "error", "description": "Method not allowed"})
